requirements.txt
# Main dependencies for QA Test Automation Framework

# Core Python testing frameworks
pytest==7.4.0
pytest-xdist==3.3.1
pytest-cov==4.1.0
behave==1.2.6

# API testing
requests==2.31.0
httpx==0.24.1

# Browser automation
playwright==1.36.0
selenium==4.10.0

# Database fixtures and factories
factory-boy==3.2.1
pytest-factoryboy==2.5.1

# Test data generation
faker==18.13.0

# Reporting
allure-pytest==2.13.2
pytest-html==3.2.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
jsonschema==4.18.4

# Database drivers
psycopg2-binary==2.9.10

# Secrets and cloud
hvac==2.2.0
boto3==1.34.40